requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
qrcode==7.4.2
pillow==10.4.0
cryptography==43.0.3
pytest==8.3.3
httpx==0.27.2
fpdf2==2.8.1
PyJWT==2.12.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# PostgreSQL support (optional for production)
psycopg2-binary==2.9.9

# Database migration and management
alembic==1.13.1
sqlalchemy==2.0.23
